Allow exit after build, take option to package
Modify the source build/install scripts to take an option
to exit after completing a build. This will leave code
behind in temp directories (why would you build and then
just throw away everything?).  This can be useful if we
want to test a build without trying to install it anywhere.

Set up the arguments to allow us to make packages instead
of directly installing the software.

+++ distro_install_scripts/Ubuntu/common/component_scripts/01_roct.sh
@@ -71,6 +71,12 @@ cd build
 cmake .. -DCMAKE_BUILD_TYPE=${ROCM_CMAKE_BUILD_TYPE} -DCMAKE_INSTALL_PREFIX=${ROCM_OUTPUT_DIR}/
 ${ROCM_SUDO_COMMAND} mkdir -p ${ROCM_OUTPUT_DIR}/lib/
 make -j `nproc`
+
+if [ ${ROCM_FORCE_BUILD_ONLY} = true ]; then
+    echo "Finished building ROCt. Exiting."
+    exit 0
+fi
+
 ${ROCM_SUDO_COMMAND} make install
 ${ROCM_SUDO_COMMAND} mkdir -p ${ROCM_OUTPUT_DIR}/include/
 ${ROCM_SUDO_COMMAND} cp -R ${SOURCE_DIR}/ROCT-Thunk-Interface/include/* ${ROCM_OUTPUT_DIR}/include/
